Solution for exercise 8.26: fix 'n+1' problem
n+1 problem: in 'allAuthors' query for each author the 'bookCount'
field has to be figured out. Which means that the code iterates over all
authors, and for each author iterates through all books, to check how many
was written by the current author. O(n*m) order of growth.
To improve the number of database queries, added a new field to the author record called
'bookCount'. It is continuously updated when a new book of the author is added to the
database. Therefore when the allAuthors response is generated, the bookCount value
is already current for each author.